backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "mcp==1.12.2",
    "openai-agents>=0.2.4",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "httpx>=0.27.0",
]
